[lib-react] Rename `sam-react-common` to `lib-react`
I want to give the builder some better heuristics to distinguish between end-project and library when generating the CI/CD configs. Using `lib-` as a prefix seems better than relying on `common` string.
