+++ .github/workflows/generated-cd-samlang-demo.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - master
     paths:
-      - sam-react-common/**
       - samlang-demo/**
       - package.json
       - yarn.lock
