+++ .github/workflows/generated-cd-samlang.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - master
     paths:
-      - sam-react-common/**
       - samlang/**
       - package.json
       - yarn.lock
