+++ .github/workflows/generated-cd-www.yml
@@ -6,7 +6,6 @@ on:
     branches:
       - master
     paths:
-      - sam-react-common/**
       - www/**
       - package.json
       - yarn.lock
