+++ .gitignore
@@ -48,10 +48,10 @@
 
 ## Project Specific
 
-### sam-react-common
-sam-react-common/**/*.js
-sam-react-common/**/*.d.ts
-sam-react-common/tsconfig.tsbuildinfo
+### lib-react
+lib-react/**/*.js
+lib-react/**/*.d.ts
+lib-react/tsconfig.tsbuildinfo
 
 # Backend
