+++ blog/package.json
@@ -13,9 +13,9 @@
     "@docusaurus/core": "^2.0.0-alpha.35",
     "@docusaurus/preset-classic": "^2.0.0-alpha.35",
     "classnames": "^2.2.6",
+    "lib-react": "0.1.0",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "sam-react-common": "0.1.0"
+    "react-dom": "^16.12.0"
   },
   "browserslist": {
     "production": [
