+++ blog/src/theme/CodeBlock/index.js
@@ -6,8 +6,8 @@
  */
 
 import React from 'react';
-import CodeBlock from 'sam-react-common/CodeBlock';
-import initialize from 'sam-react-common/language';
+import CodeBlock from 'lib-react/CodeBlock';
+import initialize from 'lib-react/language';
 
 import styles from './styles.module.css';
