+++ lib-react/README.md
@@ -1,4 +1,4 @@
-# sam-react-common
+# lib-react
 
 A collection of React components for Developer Sam's websites.
