+++ lib-react/package.json
@@ -1,11 +1,7 @@
 {
-  "name": "sam-react-common",
+  "name": "lib-react",
   "version": "0.1.0",
   "description": "The common React code used across Developer Sam's website.",
-  "keywords": [
-    "highligher",
-    "samlang"
-  ],
   "author": "Developer Sam",
   "repository": {
     "type": "git",
