+++ lib-react/tsconfig.json
