+++ samlang-demo/package.json
@@ -4,9 +4,9 @@
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.7.0",
+    "lib-react": "0.1.0",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "sam-react-common": "0.1.0"
+    "react-dom": "^16.12.0"
   },
   "scripts": {
     "start": "react-scripts start",
