+++ samlang-demo/src/LanguageDemo/ResultCard.tsx
@@ -2,8 +2,8 @@ import React, { ReactNode, ReactElement } from 'react';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardHeader from '@material-ui/core/CardHeader';
-import CodeBlock from 'sam-react-common/CodeBlock';
-import initialize from 'sam-react-common/language';
+import CodeBlock from 'lib-react/CodeBlock';
+import initialize from 'lib-react/language';
 import { Response } from './interpret';
 import DemoStyles from './LanguageDemo.module.css';
 import ResultStyles from './ResultCard.module.css';
