+++ ten/package.json
@@ -6,7 +6,7 @@
     "@material-ui/core": "^4.7.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "sam-react-common": "0.1.0"
+    "lib-react": "0.1.0"
   },
   "devDependencies": {
     "worker-loader": "^2.0.0"
