+++ ten/src/App.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import Button from '@material-ui/core/Button';
-import MaterialThemedApp from 'sam-react-common/MaterialThemedApp';
-import MaterialButtonLink from 'sam-react-common/MaterialButtonLink';
+import MaterialThemedApp from 'lib-react/MaterialThemedApp';
+import MaterialButtonLink from 'lib-react/MaterialButtonLink';
 import DistributedGameCard from './components/DistributedGameCard';
 import LocalGameCard from './components/LocalGameCard';
 import styles from './App.module.css';
