+++ www/package.json
@@ -5,9 +5,9 @@
   "dependencies": {
     "@material-ui/core": "^4.7.0",
     "@material-ui/icons": "^4.5.1",
+    "lib-react": "0.1.0",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "sam-react-common": "0.1.0"
+    "react-dom": "^16.12.0"
   },
   "scripts": {
     "start": "react-scripts start",
