+++ www/src/components/ProjectsSection/index.tsx
@@ -4,7 +4,7 @@ import CardHeader from '@material-ui/core/CardHeader';
 import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
 import Code from '@material-ui/icons/Code';
-import MaterialButtonLink from 'sam-react-common/MaterialButtonLink';
+import MaterialButtonLink from 'lib-react/MaterialButtonLink';
 import ConsoleSection from '../Common/ConsoleSection';
 import styles from './index.module.css';
 import TEN from '../../assets/projects/ten.png';
