+++ www/src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { hydrate, render } from 'react-dom';
 import './index.css';
-import initializeHighlighter from 'sam-react-common/language';
+import initializeHighlighter from 'lib-react/language';
 import App from './App';
 
 initializeHighlighter();
